% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embvr.R
\docType{package}
\name{embvr}
\alias{embvr}
\alias{embvr-package}
\title{embvr: A package for the computation of expected maximum haploid breeding values.}
\description{
This package provides a C++ routine for the fast computation of expected maximum haploid breeding values.
}
